requirements.txt
# Web Framework
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
python-dotenv==1.0.1
gunicorn==21.2.0

# Audio & Speech Processing
SpeechRecognition==3.10.1
pydub==0.25.1

# Database Drivers
psycopg2-binary==2.9.9

# Utilities
requests==2.31.0
